Split attribute page from index
The current implementation is weird, since they don't share any props or
implementation details besides using Layout.
